+++ HotelReservation/src/main/java/com/subham/hotelreservation/utils/HotelSortArgument.java
@@ -1,12 +1,12 @@
-package com.subham.hotelreservaion.service;
+package com.subham.hotelreservation.utils;
 
 import com.subham.hotelreservation.models.Hotel;
 
-public class SortArgument {
+public class HotelSortArgument {
     private Double amount;
     private Hotel h;
 
-    public SortArgument(double amount, Hotel h) {
+    public HotelSortArgument(double amount, Hotel h) {
         this.amount = amount;
         this.h = h;
     }
